requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.9.1
django-filter==2.4.0
social-auth-app-django==4.0.0
django-debug-toolbar==3.2.1
django-debug-toolbar-force==0.1.8